Handle all dependencies properly
Using go get inside the Makefile was problematic, because
poudriere grants network access only during the fetch phase.
So, doing go get inside the pre-build target was failing
when doing poudriere testport. To remedy, we fetch all
necessary files from github during the fetch phase. To do
that, we use the GH_TUPLE variable as described in the
porter's handbook chapter 5.4.3.1 (Fetching Multiple Files
from GitHub). In the pre-build phase, we move each
extracted dir to the location Go will expect to find it
when building.